- support pcie domains
Synopsis:

PCIe domain support is required in some architectures but opencl
1.2 does not supporting the GPU's pcie domain. This would cause
nvidia GPUs to be enumerated twice, once as CUDA and once as
opencl. To work around this, we do a partial match of existing
unique ids to see if the opencl type should be replaced.
This is a hack, but will work under most circumstance.